new functionality, messy code
added a laser sigh to the gun, camera smoothly moves back to the player when the scene is reloaded, and added a basic health system.
The code is quite messy right now, so I plan to refactor it before working on anything else so it doesn't spiral out of control.