New colors, new idea
I changed the color palette, which I think makes the game look much better. I was super uninterested with what I had in mind for the game, so I scrapped the linear prototype level and the enemies. I have a better idea now, so we'll see some of that in the next commit